RandomSectionクラスを別ファイルに分割
